feat: Implement neural network
Wasn't really sure what I'd go down the path with the prior code, but I got inspired by CodeBullet and some other cool projects; so decided to pivot to see if these AI boids can learn to follow the 4(5 if you deem perception to be one) rules of the model.

I have plans to add predators to see if they affect them, as well as future tests with using non-AI boids to also see if it affects growth.

Will be implementing ImGUI and potentially 3D(who knows!)